+++ superpoint/models/super_point.py
@@ -50,8 +50,8 @@ class SuperPoint(BaseModel):
             prob = tf.map_fn(lambda p: utils.box_nms(
                 p, config['nms'], keep_top_k=config['top_k']), prob)
             results['prob_nms'] = prob
-        results['pred'] = tf.to_int32(tf.greater_equal(
-            prob, config['detection_threshold']))
+        results['pred'] = tf.cast(tf.greater_equal(
+            prob, config['detection_threshold']), tf.int32)
 
         return results
